0.10.3 (#125)
* 0.10.3

Update sentry dependencies and convert date display to local format

The Sentry dependencies have been updated to a newer version. Additionally, the date display format for subscription and trial end dates has been changed to local language and convention. This ensures better alignment with the local settings of the user.