0.10.12 (#133)
Update package versions and refactor AdminUserPage

This commit updates several package versions as part of the routine maintenance and evolution of the software. Key packages that were updated include 'supabase-js', 'react-table', and 'eslint' among others to their latest versions.

The AdminUserPage function was also refactored to change a synchronous function into an asynchronous call for loading data. The organization UUID was added to the displayed table data.