Update authentication redirect and package dependencies (#138)
The import statement for redirection was adjusted along with how redirection is handled for users that are signed in and do not require MFA. Multiple package dependencies were updated to newer versions, resulting in changes in the `package.json` and `package-lock.json` files. Also, test assertions were added to ensure that signed in users are redirected away from the authentication pages.