0.10.33 2 (#157)
* 0.10.33

Improve logging and update package versions

Added more comprehensive logging to the loadAppData function to better handle and report cases where a user record is not found, the user is not yet onboarded, or the user isn't part of an organization. Also updated several dependencies, including "@supabase/supabase-js", "i18next", "@types/react", and "supabase", enhancing the application's performance and security.