0.11.2 (#162)
Add URL length constraints and upgrade package version

The commit introduces length constraints on the 'logo_url' field for the 'organizations' table and the 'photo_url' field for the 'users' table in the database. It also upgrades the 'next-supabase-saas-kit' package version from 0.11.1 to 0.11.2, along with updates to the associated '@sentry' dependencies.