0.11.13 (#172)
* 0.11.13

Refactor auth middleware and update dependencies in package-lock.json

This commit includes a significant adjustment to the authentication middleware, with even handling for 'data' and 'error' conditions. The host is now retrieved from the incoming request, modifying the redirects for sign-in and 404 situations. Also, the package-lock.json file has been updated due to dependency upgrades, improving the overall package stability and security.