Switch from getSession to getUser for Supabase auth (#178)
* Switch from getSession to getUser for Supabase auth

Replaced usage of `getSession` with `getUser` in authentication-related functions to improve accuracy and clarity. Additionally, updated various dependencies, including Supabase and Next.js, to their latest versions for enhanced performance and security.

* Remove accessToken from AuthChangeListener.

This change simplifies the AuthChangeListener component by removing the unnecessary accessToken prop. The associated components and types have been updated to reflect this change. This improves maintainability by reducing the number of dependencies within the authentication flow.

* Add 'use client' directive to RecoverStripeCheckout component

This change ensures that the 'RecoverStripeCheckout' component runs on the client side. It helps to avoid potential issues with server-side rendering.